Removed unused code, consolidated some useless splitting
